requirements.txt
alembic==1.13.1
autopep8==2.0.4
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Scss==0.5
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
Flask-WTF==1.2.1
greenlet==3.0.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mypy-extensions==1.0.0
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pycodestyle==2.11.1
pyScss==1.4.0
python-decouple==3.8
python-dotenv==1.0.1
six==1.16.0
SQLAlchemy==2.0.29
toml==0.10.2
typing_extensions==4.10.0
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.17.0
